// src/types/form.ts
export interface CompanyFormData {
  fullName: string
  email: string
  companyName: string
  alternativeCompanyName: string
  companyDesignation: string
  jurisdictionOfOperation: string
  targetJurisdictions: string
  numberOfShares: string
  areAllSharesIssued: string
  numberOfIssuedShares: string
  valuePerShares: string
}